refactor: admin/diet-sessions featureをserver/client/shared構成にリストラクチャ
