+++ admin/src/features/diet-sessions/client/components/diet-session-form.tsx
@@ -6,7 +6,7 @@ import { toast } from "sonner";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { createDietSession } from "../actions/create-diet-session";
+import { createDietSession } from "../../server/actions/create-diet-session";
 
 export function DietSessionForm() {
   const nameId = useId();
