+++ admin/src/features/diet-sessions/server/loaders/load-diet-sessions.ts
@@ -1,4 +1,4 @@
-import type { DietSession } from "../types";
+import type { DietSession } from "../../shared/types";
 import { findAllDietSessions } from "../repositories/diet-session-repository";
 
 export async function loadDietSessions(): Promise<DietSession[]> {
